refactor top-level pkg to vmware_nsx_tempest_plugin
This patch refactors the source and supporting files to use the top
level package name of vmware_nsx_tempest_plugin. This better matches
the project name as well as the name we plan to publish to PYPI with
as per the discussion in [1].

A sample release has been published to the test pypi repo [2] to
ensure this works.

[1] https://review.openstack.org/#/c/584498/
[2] https://test.pypi.org/project/vmware-nsx-tempest-plugin/

Change-Id: I4cd89f49562c780754ebfb7e93c38b4e6556e314